Port wf (#281)
https://github.com/aws-observability/aws-otel-dotnet-instrumentation/commit/ab3f8ada50de7fbc44486065fd3bdd5f1a45e8e3
This commit added gates to PR on the `release/v*` branches before it can
be merged but the required workflows are either not available in this
release branch or not up-to-date.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.