Upgraded OTEL runtime dependencies to 1.14 and OTEL AutoInstrumentation to 1.13 (#293)
*Description of changes:*

There was a request to upgrade `OpenTelemetry.Instrumentation.AWSLambda`
to `1.14` which meant that we needed to also update the
`OpenTelemetry.*` low level runtime dependencies to 1.14 as well.

It also seems like `Microsoft.Extensions.Logging` minimal version was
downgraded from 9.0.0 to 8.0.0 so had to also account for that.
https://github.com/open-telemetry/opentelemetry-dotnet-instrumentation/blob/a9448e4b641f79a2c701a3077ea056dfbaa11d65/CHANGELOG.md?plain=1#L47-L48


https://github.com/open-telemetry/opentelemetry-dotnet/blob/b0b42117a4d1b7f125633fcf488d003ba9062961/RELEASENOTES.md?plain=1#L12-L16

*Testing:*
Manually deployed a local testing app and verified that application
signals started and requests were instrumented. Also deployed a lambda
function with a custom instrumentation layer and verified that the
lambda function was also instrumented as expected.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.