Bump Instrumentation.AWS dependencies (#302)
*Issue #, if available:*
-
https://github.com/open-telemetry/opentelemetry-dotnet-contrib/releases/tag/Instrumentation.AWS-1.14.1

> Fix issue when using ADOT auto-instrumentation which includes a
version of
OpenTelemetry.Instrumentation.AWSLambda in the Lambda layer. In this
scenario
nothing is forcing the NuGet resolution to include the updated version
of
Amazon.Lambda.Core that has the
https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3410
for determining the trace id.

(https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3629)

*Description of changes:*
- Bump `OpenTelemetry.Extensions.AWS` and
`OpenTelemetry.Instrumentation.AWSLambda` to include above fix
- Downgrade `Amazon.Lambda.Core` in Lambda e2e test sample app to test
that it will be forced to be updated by ADOT as per the fix above.

*Description of changes:*
- Deployed Sample App with older version of `Amazon.Lambda.Core`,
verified it succeeded execution.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.